Prune redundant and low-value tests
Remove trivial assertions, exact-duplicate coverage, and tests strictly
subsumed by stronger siblings across Nav, Wiggle, index, and utils suites.
Merge FileDetails.test.ts into FileDetails.svelte.test.ts.
